Removing faulty assert in dx12 backend.
The purpose of the assert was to validate the a "1:1 translation" between state tracker transition and the issued barriers.

The reason why this 1:1 mapping is not correct:

Both D3D12_RESOURCE_STATE_COMMON=0 and D3D12_RESOURCE_STATE_PRESENT=0 have the same value. While the ResourceStates::Common = 0x00000001 and ResourceStates::Present = 0x00008000 differ.

This means when transitioning from Common -> Present the assumption "// otherwise there's an early-out in the beginning of this function" is not correct. m_D3DBarriers will be empty as it should be. While from the POV of the CommandListResourceStateTracker a transition is deemed to be necessary.